// src/usecase/customer/list/list.customer.dto.ts
// tslint:disable-next-line: no-empty-interface
export interface InputListCustomerDto {}

type Customer = {
  id: string;
  name: string;
  address: {
    street: string;
    number: number;
    zip: string;
    city: string;
  };
};

export interface OutputListCustomerDto {
  customers: Customer[];
}